backend/requirements.txt
fastapi
uvicorn
pandas
requests
joblib
pvlib
pytz
xgboost
scikit-learn
python-dotenv